Added log4j2.xml, updated build.xml and corrupt library and progress on Casilla
Added a new log4j2.xml (which is not included on the classpath for now, so it does not work)
Updated build.xml to work with the included libraries and also updated log4j2-core as is was a corrupted file.
Now Casilla removes boxes more accurately, but it needs some tweaking to work as required.